# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Pocassist Api",
        "title": "Pocassist Api",
        "contact": {},
        "version": "v0.4.0"
    },
    "paths": {
...
        "/api/v1/self/resetpwd/": {
            "post": {
                "description": "重置密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Reset Password",
                "parameters": [
                    {
                        "description": "旧/新密码",
                        "name": "resetpwd",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.ResetPwd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/info": {
            "get": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "获取个人信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Self",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/user/login": {
            "post": {
                "description": "登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "用户/密码",
                        "name": "auth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.Auth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/user/logout": {
            "get": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "登出",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Logout",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.Auth": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "auth.ResetPwd": {
            "type": "object",
            "required": [
                "newpassword",
                "password"
            ],
            "properties": {
                "newpassword": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "msg.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}